add TypeScript types to all component files; add TypeScript types for meals
