--- components/main-header/main-header.js
@@ -1,31 +1,0 @@
-import Link from "next/link";
-import Image from "next/image";
-import MainHeaderBackground from "./main-header-background";
-import NavLink from "./nav-link";
-import logoImg from '@/assets/logo.png';
-import styles from './main-header.module.scss';
-
-export default function MainHeader() {
-    return (
-        <>
-            <MainHeaderBackground />
-            <header className={styles.header}>
-                <Link className={styles.logo} href="/">
-                    <Image src={logoImg} alt="A plate with food on it" priority />
-                    NextLevel Food
-                </Link>
-
-                <nav className={styles.nav}>
-                    <ul>
-                        <li>
-                            <NavLink href="/meals">Browse Meals</NavLink>
-                        </li>
-                        <li>
-                            <NavLink href="/community">Foodies Community</NavLink>
-                        </li>
-                    </ul>
-                </nav>
-            </header>
-        </>    
-    );
-}
