--- components/main-header/nav-link.js
@@ -1,13 +1,0 @@
-'use client';
-
-import Link from "next/link";
-import { usePathname } from "next/navigation";
-import styles from './nav-link.module.scss';
-
-export default function NavLink({ href, children }) {
-    const path = usePathname();
-    
-    return (
-        <Link href={href} className={path.startsWith(href) ? `${styles.link} ${styles.active}` : styles.link}>{children}</Link>
-    );
-}
