--- components/meals/image-picker.js
@@ -1,66 +1,0 @@
-'use client';
-
-import { useRef, useState } from 'react';
-import styles from './image-picker.module.scss';
-import Image from 'next/image';
-
-export default function ImagePicker({ label, name }) {
-    const [pickedImage, setPickedImage] = useState();
-    const imageInput = useRef();
-
-    const handlePickClick = () => {
-        imageInput.current.click();
-    }
-
-    const handleImageChange = (e) => {
-        const file = e.target.files[0];
-
-        if (!file) {
-            setPickedImage(null);
-            return;
-        }
-
-        const fileReader = new FileReader();
-
-        fileReader.onload = () => {
-            setPickedImage(fileReader.result);
-        }
-
-        fileReader.readAsDataURL(file);
-    } 
-
-    return (
-        <div className={styles.picker}>
-            <label htmlFor={name}>{label}</label>
-            <div className={styles.controls}>
-                <div className={styles.preview}>
-                    {!pickedImage && <p>No image picked yet.</p>}
-                    {pickedImage && (
-                        <Image
-                            src={pickedImage}
-                            alt="The image selected by the user."
-                            fill
-                        />
-                    )}
-                </div>
-                <input
-                    className={styles.input}
-                    type="file"
-                    id={name}
-                    accept="image/png, image/jpeg"
-                    name={name}
-                    ref={imageInput}
-                    onChange={handleImageChange}
-                    required
-                />
-                <button
-                    className={styles.button}
-                    type="button"
-                    onClick={handlePickClick}
-                >
-                    Pick an Image
-                </button>
-            </div>
-        </div>
-    )
-}
