--- components/meals/meals-form-submit.js
@@ -1,11 +1,0 @@
-'use client';
-
-import { useFormStatus } from "react-dom";
-
-export default function MealsFormSubmit() {
-    const { pending } = useFormStatus();
-
-    return <button type="submit" disabled={pending}>
-        {pending ? "Submitting..." : "Share Meal"}
-    </button>
-}
