+++ components/meals/meals-form-submit.tsx
@@ -1,0 +1,13 @@
+'use client';
+
+import { useFormStatus } from 'react-dom';
+
+export default function MealsFormSubmit() {
+    const { pending } = useFormStatus();
+
+    return (
+        <button type="submit" disabled={pending}>
+            {pending ? "Submitting..." : "Share Meal"}
+        </button>
+    );
+}
