--- components/meals/meals-grid.js
@@ -1,12 +1,0 @@
-import MealItem from './meal-item';
-import styles from './meals-grid.module.scss';
-
-export default function MealsGrid({ meals }) {
-    return (
-        <ul className={styles.meals}>
-            {meals.map((meal) => <li key={meal.id}>
-                <MealItem {...meal} />
-            </li>)}
-        </ul>
-    )
-}
